Fix warnings in trace_impl

+++ trace/impl/include/trace_impl.h
@@ -20,10 +20,10 @@ typedef struct trace_t {
    * which will create a significant pause in the calling thread.
    */
   trace_record_nodeps_t** _lf_trace_buffer;
-  int* _lf_trace_buffer_size;
+  size_t * _lf_trace_buffer_size;
 
   /** The number of trace buffers allocated when tracing starts. */
-  int _lf_number_of_trace_buffers;
+  size_t _lf_number_of_trace_buffers;
 
   /** Marker that tracing is stopping or has stopped. */
   int _lf_trace_stop;
@@ -36,7 +36,7 @@ typedef struct trace_t {
 
   /** Table of pointers to a description of the object. */
   object_description_t _lf_trace_object_descriptions[TRACE_OBJECT_TABLE_SIZE];
-  int _lf_trace_object_descriptions_size;
+  size_t _lf_trace_object_descriptions_size;
 
   /** Indicator that the trace header information has been written to the file. */
   bool _lf_trace_header_written;
